Moves geth to a new package
